Working

+++ src/application/Model.java
@@ -1,7 +1,7 @@
 package application;
 
 public class Model {
-	public float calculate(float number1,float number2,String operator) {
+	public float calculate(long number1,long number2,String operator) {
 		switch (operator) {
 		case "+" :
 			return number1+number2;
